+++ src/main/java/bowling/LeftScoreCount.java
@@ -1,4 +1,4 @@
-package bowling.refactor;
+package bowling;
 
 public class LeftScoreCount {
