--- src/main/java/bowling/NextAddingUpScores.java
@@ -1,88 +1,0 @@
-package bowling;
-
-import java.util.*;
-import java.util.stream.Collectors;
-
-public class NextAddingUpScores {
-
-    public final static int MAX_SIZE = 2;
-    public final static int SPARE_ADDING_UP_COUNT = 1;
-
-    private final Queue<Score> scores;
-
-    private NextAddingUpScores(final Queue<Score> scores) {
-        this.scores = new LinkedList<>(scores);
-    }
-
-    public static NextAddingUpScores newInstance(final int... scoreNumbers) {
-        List<Score> scores = Arrays.stream(scoreNumbers)
-                .mapToObj(Score::of)
-                .collect(Collectors.toList());
-
-        return newInstance(scores);
-    }
-
-    public static NextAddingUpScores newInstance(final List<Score> scores) {
-        List<Score> addingUpScores = getAddingUpScores(scores);
-
-        LinkedList<Score> nextAddingUpScores = new LinkedList<>();
-        nextAddingUpScores.addAll(addingUpScores);
-
-        return new NextAddingUpScores(nextAddingUpScores);
-    }
-
-    private static List<Score> getAddingUpScores(final List<Score> scores) {
-        return scores.stream()
-                .limit(MAX_SIZE)
-                .collect(Collectors.toList());
-    }
-
-    public NextAddingUpScores update(final List<Score> scores) {
-        NextAddingUpScores updateNextAddingUpScores = new NextAddingUpScores(this.scores);
-        scores.forEach(updateNextAddingUpScores::add);
-        return updateNextAddingUpScores;
-    }
-
-    private void add(final Score score) {
-        if (scores.size() >= MAX_SIZE) {
-            scores.poll();
-        }
-        scores.add(score);
-    }
-
-    public int sumAddingUpStrikeCase() {
-        return Score.sum(new ArrayList<>(scores));
-    }
-
-    public int sumAddingUpSpareCase() {
-        List<Score> addingUpScore = scores.stream()
-                .limit(SPARE_ADDING_UP_COUNT)
-                .collect(Collectors.toList());
-
-        return Score.sum(addingUpScore);
-    }
-
-    public int size() {
-        return scores.size();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof NextAddingUpScores)) return false;
-        NextAddingUpScores that = (NextAddingUpScores) o;
-        return Objects.equals(scores, that.scores);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(scores);
-    }
-
-    @Override
-    public String toString() {
-        return "NextAddingUpScores{" +
-                "scores=" + scores +
-                '}';
-    }
-}
