+++ src/main/java/bowling/framestate/last/StrikeLastFrame.java
@@ -1,9 +1,9 @@
-package bowling.refactor.framestate.last;
+package bowling.framestate.last;
 
-import bowling.refactor.FrameScore;
-import bowling.refactor.LeftScoreCount;
-import bowling.refactor.Score;
-import bowling.refactor.framestate.State;
+import bowling.FrameScore;
+import bowling.LeftScoreCount;
+import bowling.Score;
+import bowling.framestate.State;
 
 import java.util.Collections;
