+++ src/test/java/bowling/refactor/FrameScoreTests.java
@@ -1,5 +1,8 @@
 package bowling.refactor;
 
+import bowling.FrameScore;
+import bowling.LeftScoreCount;
+import bowling.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
