+++ src/test/java/bowling/refactor/LeftScoreCountTests.java
@@ -1,5 +1,6 @@
 package bowling.refactor;
 
+import bowling.LeftScoreCount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
